[update]minigame

+++ Xilinx/workspace/Arty_z7_20/HDMI/HDMI_VGA/HDMI_DMA.vitis/RPG_System_ver1/src/system/wrapper/conf_wrapper.h
@@ -95,7 +95,7 @@ typedef struct event
 typedef struct display
 {
     uint32_t system;
-    uint8_t  sub_state;
+    uint32_t sub_state;
     uint8_t  drawtype;
     uint32_t uncommon_window;
 } DisplayConf;
